# api/internal/openapi/swagger.yaml
openapi: 3.0.3
info:
  title: Matching Engine API
  version: 0.1.0
  description: >-
    REST + SSE API for the matching engine. Authenticated routes use a bearer
    token (`Authorization: Bearer <token>`) obtained from `POST /sessions`.
    A session has a scope (`read` or `write`) and an origin (`login` or
    `minted`): write-scoped routes (order create/cancel) require a
    write-scoped session, and session-management routes that mint or revoke
    tokens require a login-origin session — a minted token can never expand
    its own access.
servers:
  - url: /api/v1
paths:
...
  /markets:
    get:
      operationId: getMarkets
      summary: List markets
      tags: [markets]
      security: []
      responses:
        "200":
          description: All markets.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GetMarketResponse" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /markets/{market}:
    get:
      operationId: getMarket
      summary: Get a market by ref (e.g. ETH-USDT)
      tags: [markets]
      security: []
      parameters:
        - name: market
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The market.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GetMarketResponse" }
        "404":
          description: Market not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Token issued by POST /sessions (login) or POST /sessions/tokens
        (minted, scoped). Send as `Authorization: Bearer <token>`.
  responses:
...
    InternalError:
      description: Internal server error. No internal detail is included in the body.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
...
    GetMarketResponse:
      type: object
      description: price_quantum/amount_quantum/min_order_size/max_order_size are in the market's quantum units.
      properties:
        base_symbol: { type: string }
        quote_symbol: { type: string }
        price_quantum: { type: integer, format: int64, minimum: 0 }
        amount_quantum: { type: integer, format: int64, minimum: 0 }
        min_order_size: { type: integer, format: int64, minimum: 0 }
        max_order_size: { type: integer, format: int64, minimum: 0 }
      required: [base_symbol, quote_symbol, price_quantum, amount_quantum, min_order_size, max_order_size]
...